Extract google places laundry fetch
