+++ src/components/localMap/LocalMap.jsx
@@ -8,6 +8,7 @@ import MarkerWithInfoWindow from '../markerWithInfoWindow/MarkerWithInfoWindow.j
 import LoadingDisplay from '../loadingDisplay/LoadingDisplay.jsx';
 import getUserLocation from '../../getUserLocation';
 import reverseGeoCode from '../../reverseGeoCode.js';
+import getPlacesLaundry from '../../getPlacesLaundry.js';
 import Autocomplete from '../autocomplete/Autocomplete.jsx';
 import LocationButton from '../locationBtn/locationBtn.jsx';
 
@@ -27,9 +28,6 @@ function LocalMap() {
 
   const APIKey = import.meta.env.VITE_APIKEY;
   const MAPID = import.meta.env.VITE_MAPID;
-
-  const googlePlacesURL =
-    'https://places.googleapis.com/v1/places:searchNearby';
 
   // Get user zip code to determine Loopie service options
   useEffect(() => {
@@ -58,45 +56,13 @@ function LocalMap() {
   }, [position]);
 
   useEffect(() => {
-    setLaundryLoaded(false);
     const fetchLaundryServices = async () => {
-      const reqBody = {
-        includedTypes: ['laundry'],
-        maxResultCount: 10,
-        locationRestriction: {
-          circle: {
-            center: {
-              latitude: position.lat,
-              longitude: position.lng
-            },
-            radius: 5000
-          }
-        }
-      };
-
-      try {
-        const response = await fetch(googlePlacesURL, {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-            'X-Goog-Api-Key': APIKey,
-            'X-Goog-FieldMask': '*'
-          },
-          body: JSON.stringify(reqBody)
-        });
-        const data = await response.json();
-        setLaundryServices(data.places);
-        if (!response.ok || !data.places) {
-          alert(
-            "Sorry, Google couldn't find that address. We know it's annoying, and we're working on it. Please try a different address, or our autolocate button."
-          );
-          throw new Error('Network response was not ok');
-        }
-        setLaundryLoaded(true);
-      } catch (err) {
-        console.error('Error : ', err);
-      }
+      setLaundryLoaded(false);
+      const data = await getPlacesLaundry(position, 20, 5000);
+      setLaundryServices(data);
+      setLaundryLoaded(true);
     };
+
     fetchLaundryServices();
   }, [position]);
